New Relic setup again
Created a new ErrorHandler that rescues StandardError to be used instead of ExceptionHandler which rescues Exception. For this transition period both airbrake and new relic will notify on errors. Airbrake will later be removed.
For improved logging reporting the logger is being changed to be a New relic one for UAT/QA/PROD environments. Change the Development logger to just be the ruby logger because it isn't needed and doesn't play well with new relic. This moves log rotation responsibility on to the infrastructure (logrotate).
These are being done in wilbertils to allow standardizing error reporting and logging across all myfreight applications.