Runtime error handling.
When errors are deliberately raised, IE Runtime errors, we'll put something that will likely be useful to the user in the error. Unhandled errors, IE relying on StandardError, will usually be a missing method error or argument error where the error message will be referring to something to do with code. These shouldn't be shown to the user.